Move SpaceCreateDialog under Components
